rearranges keycode values, hooks-up unicode
