+++ quantum/quantum.h
@@ -13,7 +13,6 @@
   #include "audio.h"
 #endif
 #ifdef MIDI_ENABLE
-	// #include <keymap_midi.h>
 	#include <lufa.h>
 #endif
 #include "action_layer.h"
@@ -27,23 +26,36 @@ extern uint32_t default_layer_state;
 	extern uint32_t layer_state;
 #endif
 
-bool music_activated;
+#ifdef AUDIO_ENABLE
+	bool music_activated;
+#endif
+
+#ifdef UNICODE_ENABLE
+	#define UC_OSX 0
+	#define UC_LNX 1
+	#define UC_WIN 2
+	#define UC_BSD 3
+
+	void set_unicode_input_mode(uint8_t os_target);
+#endif
+
+#ifndef DISABLE_LEADER
+	void leader_start(void);
+	void leader_end(void);
+
+	#ifndef LEADER_TIMEOUT
+		#define LEADER_TIMEOUT 200
+	#endif
+	#define SEQ_ONE_KEY(key) if (leader_sequence[0] == (key) && leader_sequence[1] == 0 && leader_sequence[2] == 0)
+	#define SEQ_TWO_KEYS(key1, key2) if (leader_sequence[0] == (key1) && leader_sequence[1] == (key2) && leader_sequence[2] == 0)
+	#define SEQ_THREE_KEYS(key1, key2, key3) if (leader_sequence[0] == (key1) && leader_sequence[1] == (key2) && leader_sequence[2] == (key3))
+
+	#define LEADER_EXTERNS() extern bool leading; extern uint16_t leader_time; extern uint16_t leader_sequence[3]; extern uint8_t leader_sequence_size
+	#define LEADER_DICTIONARY() if (leading && timer_elapsed(leader_time) > LEADER_TIMEOUT)
+#endif
 
 void matrix_init_kb(void);
 void matrix_scan_kb(void);
 bool process_action_kb(keyrecord_t *record);
-
-void leader_start(void);
-void leader_end(void);
-
-#ifndef LEADER_TIMEOUT
-	#define LEADER_TIMEOUT 200
-#endif
-#define SEQ_ONE_KEY(key) if (leader_sequence[0] == (key) && leader_sequence[1] == 0 && leader_sequence[2] == 0)
-#define SEQ_TWO_KEYS(key1, key2) if (leader_sequence[0] == (key1) && leader_sequence[1] == (key2) && leader_sequence[2] == 0)
-#define SEQ_THREE_KEYS(key1, key2, key3) if (leader_sequence[0] == (key1) && leader_sequence[1] == (key2) && leader_sequence[2] == (key3))
-
-#define LEADER_EXTERNS() extern bool leading; extern uint16_t leader_time; extern uint16_t leader_sequence[3]; extern uint8_t leader_sequence_size
-#define LEADER_DICTIONARY() if (leading && timer_elapsed(leader_time) > LEADER_TIMEOUT)
 
 #endif
