--- quantum/unicode_targets.h
@@ -1,5 +1,0 @@
-#define UC_OSX 0
-#define UC_LNX 1
-#define UC_WIN 2
-#define UC_BSD 3
-
